+++ packages/terra-application-links/tests/jest/ApplicationTabs.test.jsx
@@ -7,7 +7,7 @@ import testLinksWithIconsConfig from './testLinksWithIconsConfig';
 
 // Snapshot tests
 it('should render ApplicationTabs with links and alignment', () => {
-  const wrapper = shallow(<div><ApplicationTabs links={testLinkConfig} alignment="start" /></div>);
+  const wrapper = enzyme.shallow(<div><ApplicationTabs links={testLinkConfig} alignment="start" /></div>);
   expect(wrapper).toMatchSnapshot();
 });
 
@@ -17,12 +17,12 @@ it('should render ApplicationTabs with icons', () => {
       <ApplicationTabs links={testLinksWithIconsConfig} />
     </div>
   );
-  const wrapper = shallow(subject);
+  const wrapper = enzyme.shallow(subject);
   expect(wrapper).toMatchSnapshot();
 });
 
 it('correctly applies the theme context className', () => {
-  const link = shallow(
+  const link = enzyme.shallow(
     <ThemeContextProvider theme={{ className: 'orion-fusion-theme' }}>
       <div>
         <ApplicationTabs links={testLinksWithIconsConfig} />