+++ packages/terra-data-grid/tests/jest/DataGrid.test.jsx
@@ -1,6 +1,4 @@
 import React from 'react';
-/* eslint-disable-next-line import/no-extraneous-dependencies */
-import { mountWithIntl } from 'terra-enzyme-intl';
 import { v4 as uuidv4 } from 'uuid';
 import DataGrid from '../../src/DataGrid';
 
@@ -69,7 +67,7 @@ afterAll(() => {
 
 describe('DataGrid', () => {
   it('verifies that the grid created is consistent with the rows and overflowColumns props', () => {
-    const wrapper = mountWithIntl(
+    const wrapper = enzymeIntl.mountWithIntl(
       <DataGrid
         id="test-terra-data-grid"
         pinnedColumns={dataFile.cols.slice(0, 2)}
@@ -99,7 +97,7 @@ describe('DataGrid', () => {
       expect(rowComponent.props.cells).toEqual(data.cells);
     };
 
-    const wrapper = mountWithIntl(
+    const wrapper = enzymeIntl.mountWithIntl(
       <DataGrid
         id="test-terra-data-grid"
         overflowColumns={dataFile.cols}
@@ -120,7 +118,7 @@ describe('DataGrid', () => {
   it('verifies row selection column header selection', () => {
     const mockOnRowSelectionHeaderColumnSelect = jest.fn();
 
-    const wrapper = mountWithIntl(
+    const wrapper = enzymeIntl.mountWithIntl(
       <DataGrid
         id="test-terra-data-grid"
         pinnedColumns={dataFile.cols.slice(0, 2)}
@@ -146,7 +144,7 @@ describe('DataGrid', () => {
   it('verifies onCellSelect callback is triggered when space is pressed on a masked cell', () => {
     const mockCellSelect = jest.fn();
 
-    const wrapper = mountWithIntl(
+    const wrapper = enzymeIntl.mountWithIntl(
       <DataGrid
         id="test-terra-data-grid"
         pinnedColumns={dataFile.cols.slice(0, 2)}
@@ -171,7 +169,7 @@ describe('DataGrid', () => {
   it('verifies onCellSelect callback is triggered when space is pressed on a non-selectable cell', () => {
     const mockCellSelect = jest.fn();
 
-    const wrapper = mountWithIntl(
+    const wrapper = enzymeIntl.mountWithIntl(
       <DataGrid
         id="test-terra-data-grid"
         pinnedColumns={dataFile.cols.slice(0, 2)}
@@ -198,7 +196,7 @@ describe('with pinned columns', () => {
   it('sets pinnedColumns as pinned', () => {
     const pinnedColumns = dataFile.cols.slice(0, 2);
 
-    const wrapper = mountWithIntl(
+    const wrapper = enzymeIntl.mountWithIntl(
       <DataGrid
         id="test-terra-data-grid"
         pinnedColumns={pinnedColumns}
@@ -215,7 +213,7 @@ describe('with pinned columns', () => {
   it('sets row selection column as pinned', () => {
     const pinnedColumns = dataFile.cols.slice(0, 2);
 
-    const wrapper = mountWithIntl(
+    const wrapper = enzymeIntl.mountWithIntl(
       <DataGrid
         id="test-terra-data-grid"
         pinnedColumns={pinnedColumns}
@@ -231,7 +229,7 @@ describe('with pinned columns', () => {
   });
 
   it('pins row selection column if pinnedColumns is undefined', () => {
-    const wrapper = mountWithIntl(
+    const wrapper = enzymeIntl.mountWithIntl(
       <DataGrid
         id="sdfdss"
         overflowColumns={dataFile.cols}