+++ packages/terra-time-input/src/TimeInput.jsx
@@ -14,7 +14,6 @@ import styles from './TimeInput.module.scss';
 
 import AccessibleInput from './_AccessibleInput';
 import TimeSpacer from './_TimeSpacer';
-import AccessibleValue from './_AccessibleValue';
 
 const cx = classNamesBind.bind(styles);
 
@@ -275,31 +274,16 @@ class TimeInput extends React.Component {
   handleSecondFocus(event) {
     this.handleFocus(event);
     this.setState({ secondInitialFocused: true });
-
-    // This check is _needed_ to avoid the contextual menu on mobile devices coming up every time the focus shifts.
-    if (!TimeUtil.isConsideredMobileDevice()) {
-      this.secondInput.setSelectionRange(0, this.secondInput.value.length);
-    }
   }
 
   handleMinuteFocus(event) {
     this.handleFocus(event);
     this.setState({ minuteInitialFocused: true });
-
-    // This check is _needed_ to avoid the contextual menu on mobile device coming up every time the focus shifts.
-    if (!TimeUtil.isConsideredMobileDevice()) {
-      this.minuteInput.setSelectionRange(0, this.minuteInput.value.length);
-    }
   }
 
   handleHourFocus(event) {
     this.handleFocus(event);
     this.setState({ hourInitialFocused: true });
-
-    // This check is _needed_ to avoid the contextual menu on mobile device coming up every time the focus shifts.
-    if (!TimeUtil.isConsideredMobileDevice()) {
-      this.hourInput.setSelectionRange(0, this.hourInput.value.length);
-    }
   }
 
   handleHourBlur(event) {
@@ -488,13 +472,22 @@ class TimeInput extends React.Component {
     } = this.state;
     const variant = TimeUtil.getVariantFromLocale(this.props);
 
+    if (event.currentTarget.value.length === 2 && event.key.match(/^[0-9]/g)) {
+      event.currentTarget.value = ''; // eslint-disable-line no-param-reassign
+    }
+
+    if (!event.key.match(/^[0-9]/g) && !(event.keyCode === KeyCode.KEY_BACK_SPACE || event.keyCode === KeyCode.KEY_DELETE || event.keyCode === KeyCode.KEY_TAB || event.keyCode === KeyCode.KEY_RIGHT || event.keyCode === KeyCode.KEY_LEFT)) {
+      event.preventDefault();
+    }
+
     if (event.key === 'n' || event.key === 'N') {
+      event.preventDefault();
       const currentTime = this.getCurrentTime();
       this.setTime(event, currentTime.hour, currentTime.minute, currentTime.second, currentTime.meridiem);
       return;
     }
 
-    if ((event.key === '-' || event.key === '_') && !this.props.atMinDate) {
+    if ((event.key === '-' || event.key === '_' || event.keyCode === KeyCode.KEY_DOWN) && !this.props.atMinDate) {
       const currentTimeValue = this.formatHour(hour, meridiem).concat(':', minute).concat(this.props.showSeconds ? ':'.concat(second) : '');
       if (TimeUtil.validateTime(currentTimeValue, this.props.showSeconds)) {
         if (inputType === 3 && this.props.showSeconds) {
@@ -540,7 +533,7 @@ class TimeInput extends React.Component {
       return;
     }
 
-    if ((event.key === '=' || event.key === '+') && !this.props.atMaxDate) {
+    if ((event.key === '=' || event.key === '+' || event.keyCode === KeyCode.KEY_UP) && !this.props.atMaxDate) {
       const currentTimeValue = this.formatHour(hour, meridiem).concat(':', minute).concat(this.props.showSeconds ? ':'.concat(second) : '');
       if (TimeUtil.validateTime(currentTimeValue, this.props.showSeconds)) {
         if (inputType === 3 && this.props.showSeconds) {
@@ -660,32 +653,9 @@ class TimeInput extends React.Component {
    * @param {Object} event Event object generated from the event delegation.
    */
   handleHourInputKeyDown(event) {
-    let stateValue = this.state.hour;
-    let { meridiem } = this.state;
+    const stateValue = this.state.hour;
+    const { meridiem } = this.state;
     const previousStateValue = stateValue;
-    const variant = TimeUtil.getVariantFromLocale(this.props);
-
-    if (event.keyCode === KeyCode.KEY_UP) {
-      stateValue = TimeUtil.incrementHour(stateValue, variant);
-
-      // Hitting 12 when incrementing up changes the meridiem
-      if (variant === TimeUtil.FORMAT_12_HOUR && stateValue === '12') {
-        if (meridiem === this.postMeridiem || !previousStateValue) {
-          meridiem = this.anteMeridiem;
-        } else {
-          meridiem = this.postMeridiem;
-        }
-      }
-    }
-
-    if (event.keyCode === KeyCode.KEY_DOWN) {
-      stateValue = TimeUtil.decrementHour(stateValue, variant);
-
-      // Hitting 11 when incrementing down changes the meridiem
-      if (variant === TimeUtil.FORMAT_12_HOUR && stateValue === '11') {
-        meridiem = meridiem === this.postMeridiem ? this.anteMeridiem : this.postMeridiem;
-      }
-    }
 
     if (stateValue !== previousStateValue) {
       this.handleValueChange(event, TimeUtil.inputType.HOUR, stateValue, meridiem);
@@ -704,16 +674,8 @@ class TimeInput extends React.Component {
    * @param {Object} event Event object generated from the event delegation.
    */
   handleMinuteInputKeyDown(event) {
-    let stateValue = this.state.minute;
+    const stateValue = this.state.minute;
     const previousStateValue = stateValue;
-
-    if (event.keyCode === KeyCode.KEY_UP) {
-      stateValue = TimeUtil.incrementMinute(stateValue);
-    }
-
-    if (event.keyCode === KeyCode.KEY_DOWN) {
-      stateValue = TimeUtil.decrementMinute(stateValue);
-    }
 
     if (previousStateValue !== stateValue) {
       this.handleValueChange(event, TimeUtil.inputType.MINUTE, stateValue, this.state.meridiem);
@@ -738,16 +700,8 @@ class TimeInput extends React.Component {
    * @param {Object} event Event object generated from the event delegation.
    */
   handleSecondInputKeyDown(event) {
-    let stateValue = this.state.second;
+    const stateValue = this.state.second;
     const previousStateValue = stateValue;
-
-    if (event.keyCode === KeyCode.KEY_UP) {
-      stateValue = TimeUtil.incrementSecond(stateValue);
-    }
-
-    if (event.keyCode === KeyCode.KEY_DOWN) {
-      stateValue = TimeUtil.decrementSecond(stateValue);
-    }
 
     if (previousStateValue !== stateValue) {
       this.handleValueChange(event, TimeUtil.inputType.SECOND, stateValue, this.state.meridiem);
@@ -852,50 +806,37 @@ class TimeInput extends React.Component {
 
   focusMinuteFromHour(event) {
     // If the hour is empty or the cursor is after the value, move focus to the minute input when the right arrow is pressed.
-    if (this.state.hour.length === 0 || this.state.hour.length === this.hourInput.selectionEnd) {
+    if (this.state.hour.length === 0) {
       this.minuteInput.focus();
-      this.minuteInput.setSelectionRange(0, 0);
       event.preventDefault();
     }
   }
 
-  focusHour(event) {
+  focusHour() {
     // If the cursor is at the left most position in the minute input, is empty or the cursor is before the value,
     // move focus to the hour input
 
-    if (this.minuteInput.selectionEnd === 0) {
+    if (this.state.minute.length === 0) {
       this.hourInput.focus();
-      if (this.state.hour) {
-        this.hourInput.setSelectionRange(this.state.hour.length, this.state.hour.length);
-        event.preventDefault();
-      }
     }
   }
 
-  focusSecondFromMinute(event) {
+  focusSecondFromMinute() {
     // If the minute is empty or the cursor is after the value, move focus to the meridiem.
     if ((this.state.minute.length === 0
       || this.state.minute.length === this.minuteInput.selectionEnd)
       && this.secondInput
     ) {
       this.secondInput.focus();
-      if (this.state.second) {
-        this.secondInput.setSelectionRange(0, 0);
-        event.preventDefault();
-      }
     }
   }
 
-  focusMinuteFromSecond(event) {
+  focusMinuteFromSecond() {
     // If the cursor is at the left most position in the second input, is empty or the cursor is before the value,
     // move focus to the minute input
 
-    if (this.secondInput.selectionEnd === 0) {
+    if (this.state.second.length === 0) {
       this.minuteInput.focus();
-      if (this.state.minute) {
-        this.minuteInput.setSelectionRange(this.state.minute.length, this.state.minute.length);
-        event.preventDefault();
-      }
     }
   }
 
@@ -1053,7 +994,7 @@ class TimeInput extends React.Component {
         ref={this.timeInputContainer}
         className={cx('time-input-container', theme.className)}
       >
-        <div className={timeInputClassNames} role="group" aria-label={this.a11yLabel}>
+        <div className={timeInputClassNames} role="group">
           {/*
           "Time of Birth group. Time of birth Hours input., ..."
         All of the controls should be presented as a group to assistive technologies. Then, each component also
@@ -1099,14 +1040,10 @@ class TimeInput extends React.Component {
         1 - why does it mean to have an invalid meridiem? Why is that possible? It is not supported by Aria or HTML to indicate a button is "invalid".
         2 - why won't the mobile invalid and mobile incomplete tests read the inputs as invalid or incomplete? The voiceOver rotor and the accessibility panel both show them correctly as incomplete or invalid, just the read-mode is wrong. This problem doesn't happen on the normal incomplete/invalid tests. I was testing on macOS the entire time.
           */}
-          <AccessibleValue
-            value={a11yTimeValue}
-            /**
-             * description: This will be read to screen reader users only textValue changes to a new time. We want to
-             * give the screen reader user feedback that their change to one of the controls has updated this time.
-             */
-            readThis={intl.formatMessage({ id: 'Terra.timeInput.labeledTextValue' },
+          <VisuallyHiddenText
+            text={intl.formatMessage({ id: 'Terra.timeInput.labeledTextValue' },
               { a11yLabel: this.a11yLabel, a11yTimeValue })}
+            aria-live="polite"
           />
           <input
             // Create a hidden input for storing the name and value attributes to use when submitting the form.
@@ -1153,7 +1090,7 @@ class TimeInput extends React.Component {
             {...inputAttributes}
             {...minuteAttributes}
             refCallback={(inputRef) => { this.minuteInput = inputRef; }}
-            label={intl.formatMessage({ id: 'Terra.timeInput.minutes' })}
+            label={intl.formatMessage({ id: 'Terra.timeInput.minutes' }, { a11yLabel: this.a11yLabel })}
             className={minuteClassNames}
             type="text"
             value={this.state.minute}
@@ -1175,7 +1112,7 @@ class TimeInput extends React.Component {
               {...inputAttributes}
               {...secondAttributes}
               refCallback={(inputRef) => { this.secondInput = inputRef; }}
-              label={intl.formatMessage({ id: 'Terra.timeInput.seconds' })}
+              label={intl.formatMessage({ id: 'Terra.timeInput.seconds' }, { a11yLabel: this.a11yLabel })}
               className={secondClassNames}
               type="text"
               value={this.state.second}