+++ packages/terra-framework-docs/CHANGELOG.md
@@ -4,6 +4,8 @@
 
 * Changed
   * Updated `terra-navigation-side-menu` example to use more meaningful labels.
+  * Updated the `terra-date-time-picker` example for field label.
+
 ## 1.64.0 - (February 7, 2024)
 
 * Changed